requirements.txt
# Serverless API dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
motor==3.3.2
pyjwt==2.8.0
httpx==0.25.2
pydantic==1.10.13
pymongo==4.6.0
python-dotenv==1.0.0